tests for aea combined into one file
